move EA enums to folder
