+++ apps/web/src/app/auth/emergency-access/request/emergency-access-update.request.ts
@@ -1,4 +1,4 @@
-import { EmergencyAccessType } from "../../core/enums/emergency-access-type";
+import { EmergencyAccessType } from "../enums/emergency-access-type";
 
 export class EmergencyAccessUpdateRequest {
   type: EmergencyAccessType;
